Refactor and cleanup project structure
