+++ src/main/java/net/explorviz/adapter/conversion/transformer/StructureTransformer.java
@@ -3,19 +3,15 @@ package net.explorviz.adapter.conversion.transformer;
 import io.opencensus.proto.trace.v1.Span;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import net.explorviz.adapter.conversion.converter.SpanStructureConverter;
-import net.explorviz.adapter.validation.SpanStructureSanitizer;
+import net.explorviz.adapter.service.converter.SpanStructureConverter;
+import net.explorviz.adapter.service.validation.SpanStructureSanitizer;
 import net.explorviz.avro.SpanStructure;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.kstream.Transformer;
 import org.apache.kafka.streams.processor.ProcessorContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 @ApplicationScoped
 public class StructureTransformer implements Transformer<byte[], Span, KeyValue<String, SpanStructure>> {
-
-  private final static Logger LOGGER = LoggerFactory.getLogger(StructureTransformer.class);
 
 
   private final SpanStructureSanitizer sanitizer;
@@ -36,7 +32,7 @@ public class StructureTransformer implements Transformer<byte[], Span, KeyValue<
 
   @Override
   public KeyValue<String, SpanStructure> transform(final byte[] key, final Span s) {
-    SpanStructure span = this.converter.toSpanStructure(s);
+    SpanStructure span = this.converter.fromOpenCensusSpan(s);
     span = sanitizer.sanitize(span);
     return new KeyValue<>(span.getLandscapeToken(), span);
   }
