+++ src/main/java/net/explorviz/adapter/service/converter/HashHelper.java
@@ -1,4 +1,4 @@
-package net.explorviz.adapter.conversion.converter;
+package net.explorviz.adapter.service.converter;
 
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
