+++ src/main/java/net/explorviz/adapter/service/converter/IdHelper.java
@@ -1,7 +1,6 @@
-package net.explorviz.adapter.conversion.converter;
+package net.explorviz.adapter.service.converter;
 
 import com.google.common.io.BaseEncoding;
-import com.google.protobuf.ByteString;
 
 /**
  * Helper class to convert IDs given as byte arrays to strings.
