+++ src/test/java/net/explorviz/adapter/conversion/ConversionStreamTest.java
@@ -20,16 +20,16 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
 import javax.inject.Inject;
-import net.explorviz.adapter.conversion.converter.SpanAttributes;
-import net.explorviz.adapter.conversion.converter.SpanDynamicConverter;
-import net.explorviz.adapter.conversion.converter.SpanStructureConverter;
+import net.explorviz.adapter.service.converter.SpanAttributes;
+import net.explorviz.adapter.service.converter.SpanDynamicConverter;
+import net.explorviz.adapter.service.converter.SpanStructureConverter;
 import net.explorviz.adapter.conversion.transformer.DynamicTransformer;
 import net.explorviz.adapter.conversion.transformer.StructureTransformer;
 import net.explorviz.adapter.injection.KafkaConfig;
-import net.explorviz.adapter.validation.NoOpStructureSanitizer;
-import net.explorviz.adapter.validation.SpanStructureSanitizer;
-import net.explorviz.adapter.validation.SpanValidator;
-import net.explorviz.adapter.validation.StrictValidator;
+import net.explorviz.adapter.service.validation.NoOpStructureSanitizer;
+import net.explorviz.adapter.service.validation.SpanStructureSanitizer;
+import net.explorviz.adapter.service.validation.SpanValidator;
+import net.explorviz.adapter.service.validation.StrictValidator;
 import net.explorviz.avro.SpanDynamic;
 import net.explorviz.avro.SpanStructure;
 import org.apache.kafka.common.serialization.Serdes;
