+++ src/test/java/net/explorviz/adapter/service/validation/DefaultValueSanitizerTest.java
@@ -1,4 +1,4 @@
-package net.explorviz.adapter.validation;
+package net.explorviz.adapter.service.validation;
 
 import net.explorviz.avro.SpanStructure;
 import net.explorviz.avro.Timestamp;
